+++ sdk/src/azure/core/az_http_policy.c
@@ -9,6 +9,7 @@
 #include <azure/core/az_version.h>
 #include <azure/core/internal/az_http_internal.h>
 #include <azure/core/internal/az_result_internal.h>
+#include <azure/core/internal/az_span_internal.h>
 
 #include <azure/core/_az_cfg.h>
 
@@ -42,6 +43,7 @@ AZ_NODISCARD az_result az_http_pipeline_policy_apiversion(
   return _az_http_pipeline_nextpolicy(ref_policies, ref_request, ref_response);
 }
 
+// "-1" below is to account for the null terminator at the end of the string.
 #define _az_TELEMETRY_ID_PREFIX "azsdk-c-"
 #define _az_TELEMETRY_ID_PREFIX_LENGTH (sizeof(_az_TELEMETRY_ID_PREFIX) - 1)
 
@@ -67,8 +69,7 @@ AZ_NODISCARD az_result az_http_pipeline_policy_telemetry(
     remainder = az_span_copy_u8(remainder, '/');
     remainder = az_span_copy(remainder, AZ_SPAN_FROM_STR(AZ_SDK_VERSION_STRING));
 
-    telemetry_id = az_span_slice(
-        telemetry_id, 0, (int32_t)(az_span_ptr(remainder) - az_span_ptr(telemetry_id)));
+    telemetry_id = az_span_slice(telemetry_id, 0, _az_span_diff(remainder, telemetry_id));
   }
 
   _az_RETURN_IF_FAILED(